refactor: 프로젝트 구조 리팩토링
- 전체 패키지를 logbat 하위로 이동
- logback 관련은 adapter로 이동
