+++ sdk/java/logbat-sdk/src/main/java/info/logbat/application/Logbat.java
@@ -1,0 +1,12 @@
+package info.logbat.application;
+
+import info.logbat.domain.options.LogbatOptions;
+
+public class Logbat {
+
+    private final LogbatOptions options;
+
+    public Logbat(LogbatOptions options) {
+        this.options = options;
+    }
+}
