--- sdk/java/logbat-sdk/src/main/java/info/logbat/logback/Logbat.java
@@ -1,7 +1,0 @@
-package info.logbat.logback;
-
-public class Logbat {
-
-    public Logbat() {
-    }
-}
