+++ sdk/java/logbat-sdk/src/test/java/info/logbat/domain/logbat/AppKeyTest.java
@@ -1,9 +1,10 @@
-package info.logbat.logback.domain.logbat;
+package info.logbat.domain.logbat;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import info.logbat.domain.logbat.AppKey;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
